+++ trex-container-app/build.sh
@@ -18,7 +18,7 @@ if [[ "$1" == "-h" ]] ; then
     exit 0
 fi
 
-TAG="${TAG:-v0.2.21}"
+TAG="${TAG:-v0.2.23}"
 
 CLI=${CLI:="podman"}
 ORG=${ORG:="rh-nfv-int"}
@@ -39,6 +39,6 @@ fi
 
 for item in ${LIST}; do
     IMAGE="${REGISTRY}/trex-container-${item}:${TAG}"
-    $CLI build ${item} -f ${item}/Dockerfile -t $IMAGE $EXTRA
+    $CLI build ${item} -f ${item}/Containerfile -t $IMAGE $EXTRA
     $CLI push $IMAGE
 done